[ESM] Support discarding events exceeding MaxRecordAgeInSeconds (#12531)

+++ tests/aws/services/lambda_/event_source_mapping/test_lambda_integration_sqs.py
@@ -1105,10 +1105,10 @@ class TestSQSEventSourceMapping:
                 messages_to_delete.append(
                     {"Id": message["MessageId"], "ReceiptHandle": message["ReceiptHandle"]}
                 )
-
-            aws_client.sqs.delete_message_batch(
-                QueueUrl=destination_queue_url, Entries=messages_to_delete
-            )
+            if messages_to_delete:
+                aws_client.sqs.delete_message_batch(
+                    QueueUrl=destination_queue_url, Entries=messages_to_delete
+                )
             assert sum([len(batch) for batch in batches]) == 15
             return [message for batch in batches for message in batch]
